core/requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==4.2
django-autoslug==1.9.9
django-debug-toolbar==4.2.0
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
flake8==7.0.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pytest==7.4.4
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.17.1
sqlparse==0.4.4
uritemplate==4.1.1
